add check photo file exit and better status get_photo_from_list_url_str_photo_file,() add better status get_list_url_str_photo_file()

+++ get-douban-photos.py
@@ -36,12 +36,16 @@ def get_photo_from_list_url_str_photo_file(input_list_url_str_photo):
             try:
                 pass
                 _img_name_str = _url_str.rsplit('/', 1)[1]
-                print('Download: %s' % _url_str, end=' ')
-                _response = requests.get(_url_str, headers={'User-Agent': random.choice(USER_AGENTS)})
-                # print(_response.headers)
-                if _response.status_code == requests.codes.ok:
-                    open(_img_name_str, 'wb').write(_response.content)
-                    print('%9d/%d OK' % (_index+1, len(_list)))
+                if os.path.isfile(_img_name_str) is False:
+                    print('Download: %s' % _url_str, end=' ')
+                    _response = requests.get(_url_str, headers={'User-Agent': random.choice(USER_AGENTS)})
+                    # print(_response.headers)
+                    if _response.status_code == requests.codes.ok:
+                        open(_img_name_str, 'wb').write(_response.content)
+                        print('%9d/%d OK' % (_index+1, len(_list)))
+                else:
+                    print('Download: %s' % _url_str, end=' ')
+                    print('%9d/%d Warn: file existed' % (_index+1, len(_list)))
             except Exception as e:
                 raise
     # 2017-08-29>
@@ -57,8 +61,9 @@ def get_list_url_str_photo_file(input_list_url_str_photo):
         pass
         time.sleep(random.random())
         _tmp_url = get_url_str_photo_file(_list_url_photo_page[_index])
+        print('# get photo url from: %s' % _list_url_photo_page[_index], end=' ')
         _list.append(_tmp_url)
-        # print(_index)
+        print('%9d/%d OK' % (_index+1, len(_list_url_photo_page)))
     return(_list)
     # 2017-08-29>
